Разделяет редьюсер и делает селекторы
