+++ src/index.js
@@ -3,7 +3,8 @@ import ReactDOM from 'react-dom';
 import App from './components/app/app.jsx';
 import {createStore, applyMiddleware} from 'redux';
 import {composeWithDevTools} from 'redux-devtools-extension';
-import {reducer, Operation} from './reducer.js';
+import reducer from './reducer/reducer.js';
+import {Operation} from './reducer/operation.js';
 import {Provider} from 'react-redux';
 import thunk from 'redux-thunk';
 import {createAPI} from './api.js';
