+++ src/reducer/data/selectors.js
@@ -1,0 +1,6 @@
+import NameSpace from '../name-space.js';
+
+export const getOffers = (state) => state[NameSpace.DATA].offers;
+export const getCities = (state) => state[NameSpace.DATA].cities;
+export const getReviews = (state) => state[NameSpace.DATA].reviews;
+export const getNearbyOffers = (state) => state[NameSpace.DATA].nearbyOffers;
